initial AF_XDP socket support (RX only)
Very minimal (and unfinished) AF_XDP socket support. Currently only support the
RX (recv) side, though the TX side is relatively trivial to implement now that I
know how everything hooks together. Later commits will describe the architecture
and general design. Basically, though .. There is a large chunk of memory you
register to share between the kernel and userspace via mmap. Then you pass
ownership of portions of that chunk of memory via 4 ring buffers that are used
to transfer ownership of each of these individual slices of memory. Each slice
is the size of a page, so generally 4096 bytes. That does somewhat limit the
MTU, though I'm sure there are things that can be done about that, perhaps like
dpdk where you basically make a linked list of each of these buffers to combine
them all into "one" packet, logically. TODO: There needs to be an allocator type
system to keep track of the UMEM chunks, though for now a very simple free list
queue is doing the job just fine. Some care will need to be taken to make sure
the proxy object (FrameBuf) cannot be duplicated for mutable access (but could
for immutable, if desired). It can also not end up back into one of the other
queue's shared with the kernel, while the user still holds it. This is the only
real issue that could arise.