very rough initial testing
Finally starting to get something from the architecture as I submitted 60+ SQE's
per cycle of the runtime s/t it drives the TX send's (16 per send op) in
SKB (aka. non-zero-copy mode). In zero copy mode, it shouldn't ever need these
system calls, the difference being controlled by a flag where we can quickly
check whether we need to wakeup any producer (but mostly it's for the TX ring,
less so the Fill). Need to test on actual hardware to test zero-copy mode. Huge
pages seems to work, not sure how much it helps but in theory it should. TODO:
need to re-architect this so that we aren't actually doing allocations to keep
track of the frames. Right now I do it somewhat cleanly with moves and ownership
in other queue's managed by us. In reality that's less effecient than keeping
track of a fixed set of memory. Maybe one way to make it somehwat more efficient
is to pre-allocate any queue that will grow, but more likely it'll involve
tracking indexes of a static array/vector/whatever that we just subslice into
and use borrowing exlusively. Maybe there's a better way to manage all this but
first I had to understand the ins-and-outs of how this works. Which was harder
than it originally seemed because there's a decent amount of undocumented
behavior that you can only figure out by reading the emails about patches or
source-diving. In fact source diving did the most but some of the most
frustrating ones (that only 16, per sendto/send call can be sent out the TX
ring, for example).