latest with merge fixes.. some other stuff still broken
TODO: rework implementation and cut needless code. Need to reduce allocations
during runtime. Ok to allocate up-front but shouldn't allocate during runtime
bookkeeping. Ideally need a buffer type that can be kept and held onto for an
extended lifetime, and handled separately and in whatever order the user
chooses. This may require a 'refcount' type thing, but also some mutability
constraints. For example -- it would be safe to mutate in multiple places as
long as they can't criss-cross back and forth while mutating. The single
threaded nature should make this slightly easier, and their lifetimes can be
tied to the actual umem, making it easier to keep things safe..